discord-bot/requirements.txt
discord.py[voice]==2.5.2
aiohttp==3.11.18
python-dotenv==1.1.0
PyNaCl==1.5.0
